[IMP] hr_attendance : fixing and improving of attendance onboarding
With this commit, some texts are changed to be more comprehensible.
The mobile view was reworked to avoid broken onboarding view.

task-3999189
