+++ addons/hr_attendance/static/src/components/barcode_dialog/kiosk_barcode_dialog.xml
@@ -1,19 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <templates xml:space="preserve">
     <t t-name="hr_attendance.BarcodeDialog" t-inherit="web.BarcodeDialog">
-        <xpath expr="//CropOverlay" position="before">
-                <div t-if="isFreshDb" class="alert alert-info" role="alert">
-                    <div class="mb-2">
-                        No Badge ID on your employee, do you want to set a barcode to test ?
-                    </div>
-                    <div class="input-group">
-                        <input id="manual_barcode" t-model="state.barcode" type="text" name="barcode"
-                            class="input-group-text w-25" t-ref="demobadgeID"/>
-                        <button type="button" class="btn btn-primary" t-on-click="setBadgeID">
-                            Set Badge ID
-                        </button>
-                    </div>
-                </div>
+        <xpath expr="//CropOverlay" position="replace">
+            <video t-ref="videoPreview" muted="true" autoplay="true" playsinline="true" class="w-100 h-100"/>
         </xpath>
     </t>
 </templates>
