+++ addons/hr_attendance/static/src/scss/views/hr_attendance.scss
@@ -3,3 +3,11 @@
     margin: 0% !important;
     padding: 0% !important;
 }
+
+.o_view_nocontent:has(.oe_view_nocontent_attendance_mobile) {
+    position: sticky !important;
+}
+
+.o_nocontent_help:has(.oe_view_nocontent_attendance_mobile) {
+    position: relative;
+}
