+++ addons/hr_attendance/views/hr_attendance_view.xml
@@ -7,7 +7,12 @@
         <field name="name">hr.attendance.tree</field>
         <field name="model">hr.attendance</field>
         <field name="arch" type="xml">
-            <tree string="Employee attendances" decoration-success="color == 10" decoration-danger="color == 1" sample="1" duplicate="false" js_class="attendance_list_view">
+            <tree
+                js_class="attendance_list_view"
+                string="Employee attendances"
+                decoration-success="color == 10"
+                decoration-danger="color == 1"
+                duplicate="false">
                 <field name="employee_id" widget="many2one_avatar_user"/>
                 <field name="check_in"/>
                 <field name="check_out" options="{}"/>
@@ -208,12 +213,7 @@
                 <filter string="Errors" name="errors"
                         domain="['|', ('worked_hours', '&gt;=', 16), '&amp;', ('check_out', '=', False), ('check_in', '&lt;=',  (context_today() - datetime.timedelta(days=1)).strftime('%Y-%m-%d'))]"                />
                 <separator/>
-                <filter string="Check In" name="check_in_filter" date="check_in"/>
-                <filter string="Last 7 days" name="last_week" domain="[(
-                    'check_in','&gt;=', (
-                        context_today() + datetime.timedelta(days=-7)
-                        )
-                    )]"/>
+                <filter string="Date" name="check_in_filter" date="check_in"/>
                 <filter string="Last 3 Months" invisible="1" name="last_three_months" domain="[(
                     'check_in','&gt;=', (
                         context_today() + datetime.timedelta(days=-90)
@@ -223,10 +223,8 @@
                     <filter string="Employee" name="employee" context="{'group_by': 'employee_id'}"/>
                     <filter string="Departement" name="departement" context="{'group_by': 'department_id'}"/>
                     <filter string="Manager" name="manager" context="{'group_by': 'manager_id'}"/>
-                    <filter string="Mode (In)" name="groupby_mode_in" context="{'group_by': 'in_mode'}"/>
-                    <filter string="Mode (Out)" name="groupby_mode_out" context="{'group_by': 'out_mode'}"/>
-                    <filter string="Check In" name="groupby_name" context="{'group_by': 'check_in:week'}"/>
-                    <filter string="Check Out" name="groupby_check_out" context="{'group_by': 'check_out'}"/>
+                    <filter string="Method" name="groupby_mode_in" context="{'group_by': 'in_mode'}"/>
+                    <filter string="Date" name="groupby_name" context="{'group_by': 'check_in:day'}"/>
                 </group>
             </search>
         </field>
@@ -252,10 +250,11 @@
         <field name="name">Attendances</field>
         <field name="path">attendances</field>
         <field name="res_model">hr.attendance</field>
-        <field name="view_mode">tree,form</field>
+        <field name="view_mode">tree,form,pivot,graph</field>
         <field name="context">
             {
-                "search_default_employee": 1
+                "search_default_groupby_name": 1,
+                "search_default_employee": 2
             }
         </field>
         <field name="search_view_id" ref="hr_attendance_view_filter"/>
@@ -271,7 +270,7 @@
         <field name="context">
             {
                 "search_default_groupby_name" : 1,
-                "search_default_employee": 1,
+                "search_default_employee": 2,
                 "search_default_last_three_months": 1
             }
         </field>
